Remove the v1 data upload components
