+++ dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/data/upload/DataFlusher.kt
@@ -8,35 +8,36 @@ package com.datadog.android.core.internal.data.upload
 
 import androidx.annotation.WorkerThread
 import com.datadog.android.api.InternalLogger
-import com.datadog.android.core.internal.persistence.PayloadDecoration
+import com.datadog.android.core.internal.ContextProvider
 import com.datadog.android.core.internal.persistence.file.FileMover
 import com.datadog.android.core.internal.persistence.file.FileOrchestrator
+import com.datadog.android.core.internal.persistence.file.FileReader
 import com.datadog.android.core.internal.persistence.file.batch.BatchFileReader
 import com.datadog.android.core.internal.persistence.file.existsSafe
-import com.datadog.android.core.internal.utils.join
 
 internal class DataFlusher(
+    internal val contextProvider: ContextProvider,
     internal val fileOrchestrator: FileOrchestrator,
-    internal val decoration: PayloadDecoration,
     internal val fileReader: BatchFileReader,
+    internal val metadataFileReader: FileReader<ByteArray>,
     internal val fileMover: FileMover,
-    internal val internalLogger: InternalLogger
+    private val internalLogger: InternalLogger
 ) : Flusher {
 
     @WorkerThread
     override fun flush(uploader: DataUploader) {
+        val context = contextProvider.context
+
         val toUploadFiles = fileOrchestrator.getFlushableFiles()
         toUploadFiles.forEach {
             val batch = fileReader.readData(it)
-                .map { it.data }
-                .join(
-                    separator = decoration.separatorBytes,
-                    prefix = decoration.prefixBytes,
-                    suffix = decoration.suffixBytes,
-                    internalLogger = internalLogger
-                )
-            uploader.upload(batch)
             val metaFile = fileOrchestrator.getMetadataFile(it)
+            val meta = if (metaFile != null && metaFile.existsSafe(internalLogger)) {
+                metadataFileReader.readData(metaFile)
+            } else {
+                null
+            }
+            uploader.upload(context, batch, meta)
             fileMover.delete(it)
             if (metaFile?.existsSafe(internalLogger) == true) {
                 fileMover.delete(metaFile)
