--- dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/data/upload/v2/DataUploadScheduler.kt
@@ -1,55 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.core.internal.data.upload.v2
-
-import com.datadog.android.api.InternalLogger
-import com.datadog.android.core.internal.ContextProvider
-import com.datadog.android.core.internal.configuration.DataUploadConfiguration
-import com.datadog.android.core.internal.data.upload.UploadScheduler
-import com.datadog.android.core.internal.net.info.NetworkInfoProvider
-import com.datadog.android.core.internal.persistence.Storage
-import com.datadog.android.core.internal.system.SystemInfoProvider
-import com.datadog.android.core.internal.utils.scheduleSafe
-import java.util.concurrent.ScheduledThreadPoolExecutor
-import java.util.concurrent.TimeUnit
-
-internal class DataUploadScheduler(
-    storage: Storage,
-    dataUploader: DataUploader,
-    contextProvider: ContextProvider,
-    networkInfoProvider: NetworkInfoProvider,
-    systemInfoProvider: SystemInfoProvider,
-    uploadConfiguration: DataUploadConfiguration,
-    private val scheduledThreadPoolExecutor: ScheduledThreadPoolExecutor,
-    private val internalLogger: InternalLogger
-) : UploadScheduler {
-
-    internal val runnable = DataUploadRunnable(
-        scheduledThreadPoolExecutor,
-        storage,
-        dataUploader,
-        contextProvider,
-        networkInfoProvider,
-        systemInfoProvider,
-        uploadConfiguration,
-        internalLogger = internalLogger
-    )
-
-    override fun startScheduling() {
-        scheduledThreadPoolExecutor.scheduleSafe(
-            "Data upload",
-            runnable.currentDelayIntervalMs,
-            TimeUnit.MILLISECONDS,
-            internalLogger,
-            runnable
-        )
-    }
-
-    override fun stopScheduling() {
-        scheduledThreadPoolExecutor.remove(runnable)
-    }
-}
