--- dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/data/upload/v2/Flusher.kt
@@ -1,19 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.core.internal.data.upload.v2
-
-import androidx.annotation.WorkerThread
-import com.datadog.tools.annotation.NoOpImplementation
-
-// TODO RUMM-0000 Should replace com.datadog.android.core.internal.net.Flusher once
-//  features are configured as V2
-@NoOpImplementation
-internal interface Flusher {
-
-    @WorkerThread
-    fun flush(uploader: DataUploader)
-}
