--- dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/data/upload/v2/NoOpDataUploader.kt
@@ -1,17 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.core.internal.data.upload.v2
-
-import com.datadog.android.api.context.DatadogContext
-import com.datadog.android.api.storage.RawBatchEvent
-import com.datadog.android.core.internal.data.upload.UploadStatus
-
-internal class NoOpDataUploader : DataUploader {
-    override fun upload(context: DatadogContext, batch: List<RawBatchEvent>, batchMeta: ByteArray?): UploadStatus {
-        return UploadStatus.UnknownStatus
-    }
-}
