+++ dd-sdk-android-core/src/test/kotlin/com/datadog/android/core/internal/data/upload/DataFlusherTest.kt
@@ -7,14 +7,16 @@
 package com.datadog.android.core.internal.data.upload
 
 import com.datadog.android.api.InternalLogger
+import com.datadog.android.api.context.DatadogContext
 import com.datadog.android.api.storage.RawBatchEvent
-import com.datadog.android.core.internal.persistence.PayloadDecoration
+import com.datadog.android.core.internal.ContextProvider
 import com.datadog.android.core.internal.persistence.file.FileMover
 import com.datadog.android.core.internal.persistence.file.FileOrchestrator
+import com.datadog.android.core.internal.persistence.file.FileReader
 import com.datadog.android.core.internal.persistence.file.batch.BatchFileReader
 import com.datadog.android.utils.forge.Configurator
 import fr.xgouchet.elmyr.Forge
-import fr.xgouchet.elmyr.annotation.StringForgery
+import fr.xgouchet.elmyr.annotation.Forgery
 import fr.xgouchet.elmyr.junit5.ForgeConfiguration
 import fr.xgouchet.elmyr.junit5.ForgeExtension
 import org.junit.jupiter.api.BeforeEach
@@ -47,33 +49,33 @@ internal class DataFlusherTest {
     @Mock
     lateinit var mockFileOrchestrator: FileOrchestrator
 
-    lateinit var payloadDecoration: PayloadDecoration
-
     @Mock
     lateinit var mockFileReader: BatchFileReader
+
+    @Mock
+    lateinit var mockMetaFileReader: FileReader<ByteArray>
 
     @Mock
     lateinit var mockFileMover: FileMover
 
     @Mock
+    lateinit var mockContextProvider: ContextProvider
+
+    @Mock
     lateinit var mockInternalLogger: InternalLogger
 
-    @StringForgery
-    lateinit var fakePrefix: String
-
-    @StringForgery
-    lateinit var fakeSuffix: String
-
-    @StringForgery
-    lateinit var fakeSeparator: String
+    @Forgery
+    lateinit var fakeContext: DatadogContext
 
     @BeforeEach
     fun `set up`() {
-        payloadDecoration = PayloadDecoration(fakePrefix, fakeSuffix, fakeSeparator)
+        whenever(mockContextProvider.context) doReturn fakeContext
+
         testedFlusher = DataFlusher(
+            mockContextProvider,
             mockFileOrchestrator,
-            payloadDecoration,
             mockFileReader,
+            mockMetaFileReader,
             mockFileMover,
             mockInternalLogger
         )
@@ -85,6 +87,10 @@ internal class DataFlusherTest {
     ) {
         // Given
         val fakeFiles = forge.aList { mock<File>() }
+        val fakeMetaFiles =
+            forge.aList(fakeFiles.size) {
+                forge.aNullable { mock<File>().apply { whenever(exists()) doReturn true } }
+            }
         val fakeBatches = forge
             .aList(fakeFiles.size) {
                 forge
@@ -93,24 +99,27 @@ internal class DataFlusherTest {
                     }
                     .map { RawBatchEvent(it.toByteArray()) }
             }
+        val fakeMeta = fakeMetaFiles.map { if (it != null) forge.aString().toByteArray() else null }
         whenever(mockFileOrchestrator.getFlushableFiles()).thenReturn(fakeFiles)
         fakeFiles.forEachIndexed { index, file ->
             whenever(
                 mockFileReader.readData(file)
             ).thenReturn(fakeBatches[index])
+            val fakeMetaFile = fakeMetaFiles[index]
+            whenever(
+                mockFileOrchestrator.getMetadataFile(file)
+            ).thenReturn(fakeMetaFile)
+            if (fakeMetaFile != null) {
+                whenever(mockMetaFileReader.readData(fakeMetaFile)).thenReturn(fakeMeta[index])
+            }
         }
 
         // When
         testedFlusher.flush(mockDataUploader)
 
         // Then
-        fakeBatches.forEach {
-            val expectedPayload =
-                payloadDecoration.prefixBytes + it.map { it.data }.reduce { acc, bytes ->
-                    acc + payloadDecoration.separatorBytes + bytes
-                } + payloadDecoration.suffixBytes
-
-            verify(mockDataUploader).upload(expectedPayload)
+        fakeBatches.forEachIndexed { index, batch ->
+            verify(mockDataUploader).upload(fakeContext, batch, fakeMeta[index])
         }
     }
 
@@ -151,6 +160,39 @@ internal class DataFlusherTest {
         fakeMetaFiles.forEach {
             verify(mockFileMover).delete(it)
         }
+    }
+
+    @Test
+    fun `M not attempt to read metadata W flush { meta doesn't exist }`(
+        forge: Forge
+    ) {
+        // Given
+        val fakeFiles = forge.aList { mock<File>() }
+        val fakeBatches = forge
+            .aList(fakeFiles.size) {
+                forge
+                    .aList {
+                        forge.aString()
+                    }
+                    .map { RawBatchEvent(it.toByteArray()) }
+            }
+        whenever(mockFileOrchestrator.getFlushableFiles()).thenReturn(fakeFiles)
+        fakeFiles.forEachIndexed { index, file ->
+            whenever(
+                mockFileReader.readData(file)
+            ).thenReturn(fakeBatches[index])
+            val fakeBatchFile =
+                forge.aNullable { mock<File>().apply { whenever(exists()) doReturn false } }
+            whenever(
+                mockFileOrchestrator.getMetadataFile(file)
+            ).thenReturn(fakeBatchFile)
+        }
+
+        // When
+        testedFlusher.flush(mockDataUploader)
+
+        // Then
+        verifyNoInteractions(mockMetaFileReader)
     }
 
     @Test
