+++ dd-sdk-android-core/src/test/kotlin/com/datadog/android/core/internal/data/upload/DataUploadRunnableTest.kt
@@ -7,17 +7,22 @@
 package com.datadog.android.core.internal.data.upload
 
 import com.datadog.android.api.InternalLogger
+import com.datadog.android.api.context.DatadogContext
 import com.datadog.android.api.context.NetworkInfo
-import com.datadog.android.core.configuration.UploadFrequency
+import com.datadog.android.api.storage.RawBatchEvent
+import com.datadog.android.core.internal.ContextProvider
+import com.datadog.android.core.internal.configuration.DataUploadConfiguration
 import com.datadog.android.core.internal.net.info.NetworkInfoProvider
-import com.datadog.android.core.internal.persistence.Batch
-import com.datadog.android.core.internal.persistence.DataReader
+import com.datadog.android.core.internal.persistence.BatchData
+import com.datadog.android.core.internal.persistence.BatchId
+import com.datadog.android.core.internal.persistence.Storage
 import com.datadog.android.core.internal.system.SystemInfo
 import com.datadog.android.core.internal.system.SystemInfoProvider
 import com.datadog.android.utils.forge.Configurator
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.annotation.Forgery
 import fr.xgouchet.elmyr.annotation.IntForgery
+import fr.xgouchet.elmyr.annotation.StringForgery
 import fr.xgouchet.elmyr.junit5.ForgeConfiguration
 import fr.xgouchet.elmyr.junit5.ForgeExtension
 import org.assertj.core.api.Assertions.assertThat
@@ -28,21 +33,24 @@ import org.junit.jupiter.api.extension.Extensions
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
 import org.mockito.Mock
+import org.mockito.invocation.InvocationOnMock
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.junit.jupiter.MockitoSettings
 import org.mockito.kotlin.any
-import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.doAnswer
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.eq
-import org.mockito.kotlin.never
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.reset
 import org.mockito.kotlin.same
 import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoInteractions
+import org.mockito.kotlin.verifyNoMoreInteractions
 import org.mockito.kotlin.whenever
 import org.mockito.quality.Strictness
+import org.mockito.stubbing.Answer
 import java.util.concurrent.ScheduledThreadPoolExecutor
 import java.util.concurrent.TimeUnit
 
@@ -58,7 +66,7 @@ internal class DataUploadRunnableTest {
     lateinit var mockThreadPoolExecutor: ScheduledThreadPoolExecutor
 
     @Mock
-    lateinit var mockReader: DataReader
+    lateinit var mockStorage: Storage
 
     @Mock
     lateinit var mockDataUploader: DataUploader
@@ -70,18 +78,25 @@ internal class DataUploadRunnableTest {
     lateinit var mockSystemInfoProvider: SystemInfoProvider
 
     @Mock
+    lateinit var mockContextProvider: ContextProvider
+
+    @Mock
     lateinit var mockInternalLogger: InternalLogger
 
     @Forgery
-    lateinit var fakeUploadFrequency: UploadFrequency
-
-    lateinit var testedRunnable: DataUploadRunnable
+    lateinit var fakeContext: DatadogContext
 
     @Forgery
-    lateinit var fakeUploadSuccess: UploadStatus.Success
+    lateinit var fakeDataUploadConfiguration: DataUploadConfiguration
+
+    private var expectedBatchesHandled: Int = 0
+
+    private lateinit var testedRunnable: DataUploadRunnable
 
     @BeforeEach
     fun `set up`(forge: Forge) {
+        // to make sure the existing tests based only on 1 batch are not broken
+        fakeDataUploadConfiguration = fakeDataUploadConfiguration.copy(maxBatchesPerUploadJob = 1)
         val fakeNetworkInfo =
             NetworkInfo(
                 forge.aValueFrom(
@@ -89,6 +104,7 @@ internal class DataUploadRunnableTest {
                     exclude = listOf(NetworkInfo.Connectivity.NETWORK_NOT_CONNECTED)
                 )
             )
+        expectedBatchesHandled = fakeDataUploadConfiguration.maxBatchesPerUploadJob
         whenever(mockNetworkInfoProvider.getLatestNetworkInfo()) doReturn fakeNetworkInfo
         val fakeSystemInfo = SystemInfo(
             batteryFullOrCharging = true,
@@ -98,32 +114,33 @@ internal class DataUploadRunnableTest {
         )
         whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
 
+        whenever(mockContextProvider.context) doReturn fakeContext
+
         testedRunnable = DataUploadRunnable(
             mockThreadPoolExecutor,
-            mockReader,
+            mockStorage,
             mockDataUploader,
+            mockContextProvider,
             mockNetworkInfoProvider,
             mockSystemInfoProvider,
-            fakeUploadFrequency,
+            fakeDataUploadConfiguration,
             mockInternalLogger
         )
     }
 
     @Test
-    fun `doesn't send batch when offline`(@Forgery batch: Batch) {
-        val networkInfo =
-            NetworkInfo(
-                NetworkInfo.Connectivity.NETWORK_NOT_CONNECTED
-            )
+    fun `doesn't send batch when offline`() {
+        // Given
+        val networkInfo = NetworkInfo(
+            NetworkInfo.Connectivity.NETWORK_NOT_CONNECTED
+        )
         whenever(mockNetworkInfoProvider.getLatestNetworkInfo()) doReturn networkInfo
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
 
+        // When
         testedRunnable.run()
 
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader, never()).upload(batch.data)
+        // Then
+        verifyNoInteractions(mockDataUploader, mockStorage)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -133,24 +150,43 @@ internal class DataUploadRunnableTest {
 
     @Test
     fun `M send batch W run() { batteryFullOrCharging }`(
-        @Forgery batch: Batch,
-        @IntForgery(min = 0, max = DataUploadRunnable.LOW_BATTERY_THRESHOLD) batteryLevel: Int
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        @IntForgery(min = 0, max = DataUploadRunnable.LOW_BATTERY_THRESHOLD) batteryLevel: Int,
+        forge: Forge
     ) {
+        // Given
         val fakeSystemInfo = SystemInfo(
             batteryFullOrCharging = true,
             batteryLevel = batteryLevel,
             onExternalPowerSource = false,
             powerSaveMode = false
         )
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
 
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
+
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
+
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn forge.getForgery(UploadStatus.Success::class.java)
+
+        // When
         testedRunnable.run()
 
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
+        // Then
+        verify(mockStorage, times(expectedBatchesHandled)).confirmBatchRead(
+            eq(batchId),
+            any(),
+            eq(true)
+        )
+        verify(mockDataUploader, times(expectedBatchesHandled)).upload(fakeContext, batch, batchMetadata)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -160,24 +196,43 @@ internal class DataUploadRunnableTest {
 
     @Test
     fun `M send batch W run() { battery level high }`(
-        @Forgery batch: Batch,
-        @IntForgery(min = DataUploadRunnable.LOW_BATTERY_THRESHOLD + 1) batteryLevel: Int
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        @IntForgery(min = DataUploadRunnable.LOW_BATTERY_THRESHOLD + 1) batteryLevel: Int,
+        forge: Forge
     ) {
+        // Given
         val fakeSystemInfo = SystemInfo(
             batteryLevel = batteryLevel,
             batteryFullOrCharging = false,
             onExternalPowerSource = false,
             powerSaveMode = false
         )
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
 
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
+
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn forge.getForgery(UploadStatus.Success::class.java)
+
+        // When
         testedRunnable.run()
 
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
+        // Then
+        verify(mockStorage, times(expectedBatchesHandled)).confirmBatchRead(
+            eq(batchId),
+            any(),
+            eq(true)
+        )
+        verify(mockDataUploader, times(expectedBatchesHandled))
+            .upload(fakeContext, batch, batchMetadata)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -187,8 +242,10 @@ internal class DataUploadRunnableTest {
 
     @Test
     fun `M send batch W run() { onExternalPower }`(
-        @Forgery batch: Batch,
-        @IntForgery(min = 0, max = DataUploadRunnable.LOW_BATTERY_THRESHOLD) batteryLevel: Int
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        @IntForgery(min = 0, max = DataUploadRunnable.LOW_BATTERY_THRESHOLD) batteryLevel: Int,
+        forge: Forge
     ) {
         val fakeSystemInfo = SystemInfo(
             onExternalPowerSource = true,
@@ -196,15 +253,31 @@ internal class DataUploadRunnableTest {
             batteryFullOrCharging = false,
             powerSaveMode = false
         )
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
 
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
+
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn forge.getForgery(UploadStatus.Success::class.java)
+
+        // When
         testedRunnable.run()
 
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
+        // Then
+        verify(mockStorage, times(expectedBatchesHandled)).confirmBatchRead(
+            eq(batchId),
+            any(),
+            eq(true)
+        )
+        verify(mockDataUploader, times(expectedBatchesHandled))
+            .upload(fakeContext, batch, batchMetadata)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -214,24 +287,22 @@ internal class DataUploadRunnableTest {
 
     @Test
     fun `M not send batch W run() { not enough battery }`(
-        @Forgery batch: Batch,
         @IntForgery(min = 0, max = DataUploadRunnable.LOW_BATTERY_THRESHOLD) batteryLevel: Int
     ) {
+        // Given
         val fakeSystemInfo = SystemInfo(
             batteryLevel = batteryLevel,
             batteryFullOrCharging = false,
             onExternalPowerSource = false,
             powerSaveMode = false
         )
-        whenever(mockReader.lockAndReadNext()) doReturn batch
         whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
 
+        // When
         testedRunnable.run()
 
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader, never()).upload(batch.data)
+        // Then
+        verifyNoInteractions(mockStorage, mockDataUploader)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -240,25 +311,23 @@ internal class DataUploadRunnableTest {
     }
 
     @Test
-    fun `M not send batch W run() { batteryFullOrCharging, powerSaveMode}`(
-        @Forgery batch: Batch,
+    fun `M not send batch W run() { batteryFullOrCharging, powerSaveMode }`(
         @IntForgery(min = 0, max = 100) batteryLevel: Int
     ) {
+        // Given
         val fakeSystemInfo = SystemInfo(
             batteryFullOrCharging = true,
             powerSaveMode = true,
             batteryLevel = batteryLevel,
             onExternalPowerSource = false
         )
-        whenever(mockReader.lockAndReadNext()) doReturn batch
         whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
 
+        // When
         testedRunnable.run()
 
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader, never()).upload(batch.data)
+        // Then
+        verifyNoInteractions(mockStorage, mockDataUploader)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -267,25 +336,23 @@ internal class DataUploadRunnableTest {
     }
 
     @Test
-    fun `M not send batch W run() { batteryLeveHigh, powerSaveMode}`(
-        @Forgery batch: Batch,
+    fun `M not send batch W run() { batteryLeveHigh, powerSaveMode }`(
         @IntForgery(min = DataUploadRunnable.LOW_BATTERY_THRESHOLD + 1) batteryLevel: Int
     ) {
+        // Given
         val fakeSystemInfo = SystemInfo(
             batteryLevel = batteryLevel,
             powerSaveMode = true,
             batteryFullOrCharging = false,
             onExternalPowerSource = false
         )
-        whenever(mockReader.lockAndReadNext()) doReturn batch
         whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
 
+        // When
         testedRunnable.run()
 
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader, never()).upload(batch.data)
+        // Then
+        verifyNoInteractions(mockStorage, mockDataUploader)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -294,25 +361,23 @@ internal class DataUploadRunnableTest {
     }
 
     @Test
-    fun `M not send batch W run() { onExternalPower, powerSaveMode}`(
-        @Forgery batch: Batch,
+    fun `M not send batch W run() { onExternalPower, powerSaveMode }`(
         @IntForgery(min = 0, max = DataUploadRunnable.LOW_BATTERY_THRESHOLD) batteryLevel: Int
     ) {
+        // Given
         val fakeSystemInfo = SystemInfo(
             onExternalPowerSource = true,
             powerSaveMode = true,
             batteryLevel = batteryLevel,
             batteryFullOrCharging = false
         )
-        whenever(mockReader.lockAndReadNext()) doReturn batch
         whenever(mockSystemInfoProvider.getLatestSystemInfo()) doReturn fakeSystemInfo
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
 
+        // When
         testedRunnable.run()
 
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader, never()).upload(batch.data)
+        // Then
+        verifyNoInteractions(mockStorage, mockDataUploader)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -322,12 +387,15 @@ internal class DataUploadRunnableTest {
 
     @Test
     fun `𝕄 do nothing 𝕎 no batch to send`() {
-        whenever(mockReader.lockAndReadNext()) doReturn null
+        // Given
+        whenever(mockStorage.readNextBatch()).thenReturn(null)
 
+        // When
         testedRunnable.run()
 
-        verify(mockReader, never()).drop(anyOrNull())
-        verify(mockReader, never()).release(anyOrNull())
+        // Then
+        verify(mockStorage).readNextBatch()
+        verifyNoMoreInteractions(mockStorage)
         verifyNoInteractions(mockDataUploader)
         verify(mockThreadPoolExecutor).schedule(
             eq(testedRunnable),
@@ -337,35 +405,35 @@ internal class DataUploadRunnableTest {
     }
 
     @Test
-    fun `batch sent successfully`(@Forgery batch: Batch) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn fakeUploadSuccess
-
-        testedRunnable.run()
-
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
-        verify(mockThreadPoolExecutor).schedule(
-            same(testedRunnable),
-            any(),
-            eq(TimeUnit.MILLISECONDS)
-        )
-    }
-
-    @Test
-    fun `batch kept on Network Error`(
-        @Forgery batch: Batch,
-        @Forgery status: UploadStatus.NetworkError
+    fun `batch sent successfully`(
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        forge: Forge
     ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
+        // Given
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
 
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn forge.getForgery(UploadStatus.Success::class.java)
+
+        // When
         testedRunnable.run()
 
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader).release(batch)
-        verify(mockDataUploader).upload(batch.data)
+        // Then
+        verify(mockStorage, times(expectedBatchesHandled)).confirmBatchRead(
+            any(),
+            any(),
+            eq(true)
+        )
+        verify(mockDataUploader, times(expectedBatchesHandled))
+            .upload(fakeContext, batch, batchMetadata)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -373,21 +441,80 @@ internal class DataUploadRunnableTest {
         )
     }
 
-    @Test
-    fun `batch kept after n Network Error`(
-        @Forgery batch: Batch,
+    @ParameterizedTest
+    @MethodSource("retryBatchStatusValues")
+    fun `batch kept on error`(
+        uploadStatus: UploadStatus,
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        forge: Forge
+    ) {
+        // Given
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
+
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn uploadStatus
+
+        // When
+        testedRunnable.run()
+
+        // Then
+        verify(mockStorage, times(expectedBatchesHandled)).confirmBatchRead(
+            eq(batchId),
+            any(),
+            eq(false)
+        )
+        verify(mockDataUploader, times(expectedBatchesHandled))
+            .upload(fakeContext, batch, batchMetadata)
+        verify(mockThreadPoolExecutor).schedule(
+            same(testedRunnable),
+            any(),
+            eq(TimeUnit.MILLISECONDS)
+        )
+    }
+
+    @ParameterizedTest
+    @MethodSource("retryBatchStatusValues")
+    fun `batch kept after n errors`(
+        uploadStatus: UploadStatus,
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
         @IntForgery(min = 3, max = 42) runCount: Int,
-        @Forgery status: UploadStatus.NetworkError
+        forge: Forge
     ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
+        // Given
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
 
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn uploadStatus
+
+        // WHen
         repeat(runCount) {
             testedRunnable.run()
         }
-        verify(mockDataUploader, times(runCount)).upload(batch.data)
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader, times(runCount)).release(batch)
+
+        // Then
+        verify(mockStorage, times(runCount)).confirmBatchRead(
+            eq(batchId),
+            any(),
+            eq(false)
+        )
+        verify(mockDataUploader, times(runCount * expectedBatchesHandled))
+            .upload(fakeContext, batch, batchMetadata)
         verify(mockThreadPoolExecutor, times(runCount)).schedule(
             same(testedRunnable),
             any(),
@@ -395,119 +522,38 @@ internal class DataUploadRunnableTest {
         )
     }
 
-    @Test
-    fun `batch dropped on Redirection`(
-        @Forgery batch: Batch,
-        @Forgery status: UploadStatus.HttpRedirection
+    @ParameterizedTest
+    @MethodSource("dropBatchStatusValues")
+    fun `batch dropped on error`(
+        uploadStatus: UploadStatus,
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        forge: Forge
     ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
+        // Given
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
 
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn uploadStatus
+
+        // When
         testedRunnable.run()
 
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
-        verify(mockThreadPoolExecutor).schedule(
-            same(testedRunnable),
+        // Then
+        verify(mockStorage, times(expectedBatchesHandled)).confirmBatchRead(
+            eq(batchId),
             any(),
-            eq(TimeUnit.MILLISECONDS)
+            eq(true)
         )
-    }
-
-    @Test
-    fun `batch dropped on Client Error`(
-        @Forgery batch: Batch,
-        @Forgery status: UploadStatus.HttpClientError
-    ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
-        testedRunnable.run()
-
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
-        verify(mockThreadPoolExecutor).schedule(
-            same(testedRunnable),
-            any(),
-            eq(TimeUnit.MILLISECONDS)
-        )
-    }
-
-    @Test
-    fun `batch dropped on Invalid Token Error`(
-        @Forgery batch: Batch,
-        @Forgery status: UploadStatus.InvalidTokenError
-    ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
-        testedRunnable.run()
-
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
-        verify(mockThreadPoolExecutor).schedule(
-            same(testedRunnable),
-            any(),
-            eq(TimeUnit.MILLISECONDS)
-        )
-    }
-
-    @Test
-    fun `batch kept on Server Error`(
-        @Forgery batch: Batch,
-        @Forgery status: UploadStatus.HttpServerError
-    ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
-        testedRunnable.run()
-
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader).release(batch)
-        verify(mockDataUploader).upload(batch.data)
-        verify(mockThreadPoolExecutor).schedule(
-            same(testedRunnable),
-            any(),
-            eq(TimeUnit.MILLISECONDS)
-        )
-    }
-
-    @Test
-    fun `batch kept after n Server Error`(
-        @Forgery batch: Batch,
-        @IntForgery(min = 3, max = 42) runCount: Int,
-        @Forgery status: UploadStatus.HttpServerError
-    ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
-
-        repeat(runCount) {
-            testedRunnable.run()
-        }
-
-        verify(mockDataUploader, times(runCount)).upload(batch.data)
-        verify(mockReader, never()).drop(batch)
-        verify(mockReader, times(runCount)).release(batch)
-        verify(mockThreadPoolExecutor, times(runCount)).schedule(
-            same(testedRunnable),
-            any(),
-            eq(TimeUnit.MILLISECONDS)
-        )
-    }
-
-    @Test
-    fun `batch dropped on Unknown error`(
-        @Forgery batch: Batch,
-        @Forgery status: UploadStatus.UnknownError
-    ) {
-        whenever(mockReader.lockAndReadNext()) doReturn batch
-        whenever(mockDataUploader.upload(batch.data)) doReturn status
-
-        testedRunnable.run()
-
-        verify(mockReader).drop(batch)
-        verify(mockReader, never()).release(batch)
-        verify(mockDataUploader).upload(batch.data)
+        verify(mockDataUploader, times(expectedBatchesHandled))
+            .upload(fakeContext, batch, batchMetadata)
         verify(mockThreadPoolExecutor).schedule(
             same(testedRunnable),
             any(),
@@ -517,17 +563,33 @@ internal class DataUploadRunnableTest {
 
     @Test
     fun `when has batches the upload frequency will increase`(
-        @Forgery batch: Batch
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        forge: Forge
     ) {
-        whenever(mockDataUploader.upload(any())) doReturn fakeUploadSuccess
-        whenever(mockReader.lockAndReadNext()).doReturn(batch)
+        // Given
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn forge.getForgery(UploadStatus.Success::class.java)
 
+        // When
         repeat(5) {
             testedRunnable.run()
         }
 
+        // Then
         val captor = argumentCaptor<Long>()
-        verify(mockThreadPoolExecutor, times(5))
+        verify(
+            mockThreadPoolExecutor,
+            times(5 * expectedBatchesHandled)
+        )
             .schedule(same(testedRunnable), captor.capture(), eq(TimeUnit.MILLISECONDS))
         captor.allValues.reduce { previous, next ->
             assertThat(next).isLessThan(previous)
@@ -537,12 +599,23 @@ internal class DataUploadRunnableTest {
 
     @Test
     fun `𝕄 reduce delay between runs 𝕎 upload is successful`(
-        @Forgery batch: Batch,
-        @IntForgery(16, 64) runCount: Int
+        @Forgery batch: List<RawBatchEvent>,
+        @StringForgery batchMeta: String,
+        @IntForgery(16, 64) runCount: Int,
+        forge: Forge
     ) {
         // Given
-        whenever(mockDataUploader.upload(any())) doReturn fakeUploadSuccess
-        whenever(mockReader.lockAndReadNext()).doReturn(batch)
+        val batchId = mock<BatchId>()
+        val batchMetadata = forge.aNullable { batchMeta.toByteArray() }
+
+        whenever(mockStorage.readNextBatch()).thenReturn(BatchData(batchId, batch, batchMetadata))
+        whenever(
+            mockDataUploader.upload(
+                fakeContext,
+                batch,
+                batchMetadata
+            )
+        ) doReturn forge.getForgery(UploadStatus.Success::class.java)
 
         // When
         repeat(runCount) {
@@ -571,14 +644,48 @@ internal class DataUploadRunnableTest {
     @MethodSource("dropBatchStatusValues")
     fun `𝕄 reduce delay between runs 𝕎 batch fails and should be dropped`(
         uploadStatus: UploadStatus,
-        @Forgery batch: Batch,
-        @IntForgery(16, 64) runCount: Int
+        @IntForgery(16, 64) runCount: Int,
+        forge: Forge,
+        @Forgery fakeConfiguration: DataUploadConfiguration
     ) {
         // Given
-        whenever(mockDataUploader.upload(any())) doAnswer {
-            uploadStatus
+        testedRunnable = DataUploadRunnable(
+            mockThreadPoolExecutor,
+            mockStorage,
+            mockDataUploader,
+            mockContextProvider,
+            mockNetworkInfoProvider,
+            mockSystemInfoProvider,
+            fakeConfiguration,
+            mockInternalLogger
+        )
+        // extra batches to make sure we are not reaching the limit as this will fall into the
+        // else branch and increase the interval making the test to fail
+        val batches = forge.aList(size = runCount * fakeConfiguration.maxBatchesPerUploadJob + 10) {
+            aList { getForgery<RawBatchEvent>() }
         }
-        whenever(mockReader.lockAndReadNext()).doReturn(batch)
+        val batchIds: List<BatchId> = batches.map { mock() }
+        val randomFailIndex = forge.anInt(min = 0, max = batches.size)
+        val batchMetadata = forge.aList(size = batches.size) {
+            aNullable { aString().toByteArray() }
+        }
+
+        stubStorage(batchIds, batches, batchMetadata)
+
+        batches.forEachIndexed { index, batch ->
+            val expectedStatus = if (index == randomFailIndex) {
+                uploadStatus
+            } else {
+                forge.getForgery(UploadStatus.Success::class.java)
+            }
+            whenever(
+                mockDataUploader.upload(
+                    fakeContext,
+                    batch,
+                    batchMetadata[index]
+                )
+            ) doReturn expectedStatus
+        }
 
         // When
         repeat(runCount) {
@@ -607,11 +714,8 @@ internal class DataUploadRunnableTest {
     fun `𝕄 increase delay between runs 𝕎 no batch available`(
         @IntForgery(16, 64) runCount: Int
     ) {
-        // Given
-        whenever(mockDataUploader.upload(any())) doReturn fakeUploadSuccess
-        whenever(mockReader.lockAndReadNext()) doReturn null
-
         // When
+        whenever(mockStorage.readNextBatch()) doReturn null
         repeat(runCount) {
             testedRunnable.run()
         }
@@ -638,11 +742,51 @@ internal class DataUploadRunnableTest {
     @MethodSource("retryBatchStatusValues")
     fun `𝕄 increase delay between runs 𝕎 batch fails and should be retried`(
         status: UploadStatus,
-        @IntForgery(16, 64) runCount: Int
+        @IntForgery(1, 10) runCount: Int,
+        forge: Forge,
+        @Forgery fakeConfiguration: DataUploadConfiguration
     ) {
         // Given
-        whenever(mockDataUploader.upload(any())) doReturn status
-        whenever(mockReader.lockAndReadNext()) doReturn null
+        testedRunnable = DataUploadRunnable(
+            mockThreadPoolExecutor,
+            mockStorage,
+            mockDataUploader,
+            mockContextProvider,
+            mockNetworkInfoProvider,
+            mockSystemInfoProvider,
+            fakeConfiguration,
+            mockInternalLogger
+        )
+        val batches = forge.aList(size = runCount * fakeConfiguration.maxBatchesPerUploadJob) {
+            aList { getForgery<RawBatchEvent>() }
+        }
+        val batchIds: List<BatchId> = batches.map { mock() }
+        val failIndexesSet = mutableSetOf<Int>().apply {
+            var index = 0
+            repeat(runCount) {
+                add(index)
+                index += fakeConfiguration.maxBatchesPerUploadJob
+            }
+        }
+        val batchMetadata = forge.aList(size = batches.size) {
+            aNullable { aString().toByteArray() }
+        }
+
+        stubStorage(batchIds, batches, batchMetadata)
+        batches.forEachIndexed { index, batch ->
+            val expectedStatus = if (index in failIndexesSet) {
+                status
+            } else {
+                forge.getForgery(UploadStatus.Success::class.java)
+            }
+            whenever(
+                mockDataUploader.upload(
+                    fakeContext,
+                    batch,
+                    batchMetadata[index]
+                )
+            ) doReturn expectedStatus
+        }
 
         // When
         repeat(runCount) {
@@ -667,7 +811,161 @@ internal class DataUploadRunnableTest {
         }
     }
 
+    // region maxBatchesPerJob
+
+    @Test
+    fun `𝕄 handle the maxBatchesPerJob W run{maxBatchesPerJob smaller availableBatches}`(
+        forge: Forge,
+        @Forgery fakeConfiguration: DataUploadConfiguration
+    ) {
+        // Given
+        testedRunnable = DataUploadRunnable(
+            mockThreadPoolExecutor,
+            mockStorage,
+            mockDataUploader,
+            mockContextProvider,
+            mockNetworkInfoProvider,
+            mockSystemInfoProvider,
+            fakeConfiguration,
+            mockInternalLogger
+        )
+        val batches = forge.aList(
+            size = forge.anInt(
+                min = fakeConfiguration.maxBatchesPerUploadJob + 1,
+                max = fakeConfiguration.maxBatchesPerUploadJob + 1000
+            )
+        ) {
+            aList { getForgery<RawBatchEvent>() }
+        }
+        val batchIds: List<BatchId> = batches.map { mock() }
+        val batchMetadata = forge.aList(size = batches.size) { aNullable { aString().toByteArray() } }
+        stubStorage(batchIds, batches, batchMetadata)
+        batches.forEachIndexed { index, batch ->
+            whenever(
+                mockDataUploader.upload(
+                    fakeContext,
+                    batch,
+                    batchMetadata[index]
+                )
+            ) doReturn forge.getForgery(UploadStatus.Success::class.java)
+        }
+
+        // When
+        testedRunnable.run()
+
+        // Then
+        batches.take(fakeConfiguration.maxBatchesPerUploadJob).forEachIndexed { index, batch ->
+            verify(mockDataUploader).upload(fakeContext, batch, batchMetadata[index])
+            verify(mockStorage).confirmBatchRead(
+                eq(batchIds[index]),
+                any(),
+                eq(true)
+            )
+        }
+        verifyNoMoreInteractions(mockDataUploader)
+        verify(mockThreadPoolExecutor).schedule(
+            same(testedRunnable),
+            any(),
+            eq(TimeUnit.MILLISECONDS)
+        )
+    }
+
+    @Test
+    fun `𝕄 exhaust the available batches W run {maxBatchesPerJob higher or equal availableBatches}`(
+        forge: Forge,
+        @Forgery fakeConfiguration: DataUploadConfiguration
+    ) {
+        // Given
+        testedRunnable = DataUploadRunnable(
+            mockThreadPoolExecutor,
+            mockStorage,
+            mockDataUploader,
+            mockContextProvider,
+            mockNetworkInfoProvider,
+            mockSystemInfoProvider,
+            fakeConfiguration,
+            mockInternalLogger
+        )
+        val fakeBatchesCount = forge.anInt(
+            min = 1,
+            max = fakeConfiguration.maxBatchesPerUploadJob + 1
+        )
+        val batches = forge.aList(
+            size = fakeBatchesCount
+        ) { aList { getForgery<RawBatchEvent>() } }
+        val batchIds: List<BatchId> = batches.map { mock() }
+        val batchMetadata = forge.aList(size = batches.size) { aNullable { aString().toByteArray() } }
+        stubStorage(batchIds, batches, batchMetadata)
+        batches.forEachIndexed { index, batch ->
+            whenever(
+                mockDataUploader.upload(
+                    fakeContext,
+                    batch,
+                    batchMetadata[index]
+                )
+            ) doReturn forge.getForgery(UploadStatus.Success::class.java)
+        }
+
+        // When
+        testedRunnable.run()
+
+        // Then
+        batches.forEachIndexed { index, batch ->
+            verify(mockDataUploader).upload(fakeContext, batch, batchMetadata[index])
+            verify(mockStorage).confirmBatchRead(
+                eq(batchIds[index]),
+                any(),
+                eq(true)
+            )
+        }
+        verifyNoMoreInteractions(mockDataUploader)
+        verify(mockThreadPoolExecutor).schedule(
+            same(testedRunnable),
+            any(),
+            eq(TimeUnit.MILLISECONDS)
+        )
+    }
+
+    // region Internal
+
+    private fun stubStorage(
+        batchIds: List<BatchId>,
+        batches: List<List<RawBatchEvent>>,
+        batchMeta: List<ByteArray?>
+    ) {
+        reset(mockStorage)
+        whenever(mockStorage.readNextBatch()) doAnswer object : Answer<BatchData?> {
+            var count = 0
+
+            override fun answer(invocation: InvocationOnMock): BatchData? {
+                val data = if (count >= batches.size) {
+                    null
+                } else {
+                    val batchData = BatchData(batchIds[count], batches[count], batchMeta[count])
+                    batchData
+                }
+                count++
+                return data
+            }
+        }
+    }
+
+    // endregion
+
     companion object {
+
+        @JvmStatic
+        fun retryBatchStatusValues(): List<UploadStatus> {
+            val forge = Forge().apply {
+                Configurator().configure(this)
+            }
+
+            return listOf(
+                forge.getForgery(UploadStatus.HttpServerError::class.java),
+                forge.getForgery(UploadStatus.HttpClientRateLimiting::class.java),
+                forge.getForgery(UploadStatus.NetworkError::class.java)
+            )
+        }
 
         @JvmStatic
         fun dropBatchStatusValues(): List<UploadStatus> {
@@ -680,19 +978,6 @@ internal class DataUploadRunnableTest {
                 forge.getForgery(UploadStatus.HttpRedirection::class.java),
                 forge.getForgery(UploadStatus.UnknownError::class.java),
                 forge.getForgery(UploadStatus.UnknownStatus::class.java)
-            )
-        }
-
-        @JvmStatic
-        fun retryBatchStatusValues(): List<UploadStatus> {
-            val forge = Forge().apply {
-                Configurator().configure(this)
-            }
-
-            return listOf(
-                forge.getForgery(UploadStatus.HttpServerError::class.java),
-                forge.getForgery(UploadStatus.HttpClientRateLimiting::class.java),
-                forge.getForgery(UploadStatus.NetworkError::class.java)
             )
         }
     }
